Deleted all LoggedIn activities and related files, replaced them with the dashboard activity. Also added linkage between the rat data button and the rat data list activity so the button performs as expected
