+++ m4/app/src/main/AndroidManifest.xml
@@ -18,7 +18,8 @@
         </activity>
         <activity android:name=".controllers.LoginActivity"></activity>
         <activity android:name=".controllers.RegisterActivity"></activity>
-        <activity android:name=".controllers.LoggedinActivity"></activity>
+        <activity android:name=".controllers.RatDataListActivity"></activity>
+        <activity android:name=".controllers.DashboardActivity"></activity>
     </application>
 
 </manifest>
