+++ m4/app/src/main/java/edu/gatech/cs2340/thericks/controllers/LoginActivity.java
@@ -48,7 +48,7 @@ public class LoginActivity extends AppCompatActivity {
                     if (Security.checkPassword(enteredPassword, u.getLogin())) {
                         Log.d("Login", "Successfully logged into user account: " + u.getLogin().getUsername());
                         Context context = v.getContext();
-                        Intent intent = new Intent(context, LoggedinActivity.class);
+                        Intent intent = new Intent(context, DashboardActivity.class);
                         intent.putExtra("edu.gatech.cs2340.thericks.User", u);
                         context.startActivity(intent);
                     } else {
