+++ m4/app/src/main/res/layout/activity_dashboard.xml
@@ -14,7 +14,7 @@
         app:layout_constraintTop_toTopOf="parent" />
 
     <Button
-        android:id="@+id/rat_data_button"
+        android:id="@+id/rat_data_list_button"
         android:layout_width="150dp"
         android:layout_height="wrap_content"
         android:text="@string/rat_list_data_message"
@@ -25,7 +25,7 @@
         android:layout_width="150dp"
         android:layout_height="wrap_content"
         android:text="@string/profile_message"
-        app:layout_constraintTop_toBottomOf="@id/rat_data_button" />
+        app:layout_constraintTop_toBottomOf="@id/rat_data_list_button" />
 
     <Button
         android:id="@+id/settings_button"
@@ -35,9 +35,17 @@
         app:layout_constraintTop_toBottomOf="@id/profile_button" />
 
     <Button
-        android:id="@+id/report_button"
+        android:id="@+id/report_rat_button"
         android:layout_width="150dp"
         android:layout_height="wrap_content"
         android:text="@string/report_message"
         app:layout_constraintTop_toBottomOf="@id/settings_button" />
+
+    <Button
+        android:id="@+id/logout_button"
+        android:layout_width="150dp"
+        android:layout_height="wrap_content"
+        android:text="@string/logout_message"
+        app:layout_constraintTop_toBottomOf="@id/report_rat_button" />
+
 </android.support.constraint.ConstraintLayout>
